Move logging of table access out from cron.Run() to separate function cron.Log_Table_Access()
This makes cron.Run() a lot cleaner.
Also don't care that the pg_stat_xact_user_tables data will contain
some values from the table accesses caused by cron.Run().
Before we read from pg_stat_xact_user_tables() two times, first before executing the cron job function
and then again, to calculate the exact delta, but I don't think it's worth the extra effort.

Also, make ErrorLogID and LogID bigints, just in case.

Also, log the ProcessID not the JobID as we can get the JobID from the ProcessID.

+++ FUNCTIONS/log_error.sql
@@ -1,10 +1,10 @@
 CREATE OR REPLACE FUNCTION cron.Log_Error(_ProcessID integer, _PgBackendPID integer, _PgErr text, _PgErrStr text, _PgState text, _PerlCallerInfo text, _RetryInSeconds numeric)
-RETURNS integer
+RETURNS bigint
 LANGUAGE plpgsql
 SET search_path TO public, pg_temp
 AS $FUNC$
 DECLARE
-_ErrorLogID integer;
+_ErrorLogID bigint;
 BEGIN
 
 INSERT INTO cron.ErrorLog ( ProcessID,  PgBackendPID,  PgErr,  PgErrStr,  PgState,  PerlCallerInfo,  RetryInSeconds)
