+++ TABLES/errorlog.sql
@@ -1,7 +1,7 @@
 CREATE TABLE cron.ErrorLog (
-ErrorLogID     serial  NOT NULL,
-ProcessID      integer NOT NULL,
-PgBackendPID   integer NOT NULL,
+ErrorLogID     bigserial NOT NULL,
+ProcessID      integer   NOT NULL,
+PgBackendPID   integer   NOT NULL,
 PgErr          text,
 PgErrStr       text,
 PgState        text,
